refactor the vectors extension for graphql
