+++ raphtory-graphql/src/data.rs
@@ -10,7 +10,6 @@ use crate::{
             filtering::{GraphAccessFilter, GraphRowFilter, HiddenKeys},
             namespace::Namespace,
             namespaced_item::NamespacedItem,
-            vectorised_graph::GqlVectorisedGraph,
         },
     },
     paths::{
@@ -20,14 +19,8 @@ use crate::{
     rayon::blocking_compute,
     GQLError,
 };
-
 use async_graphql::Context;
 use dynamic_graphql::Enum;
-#[cfg(feature = "vectors")]
-use raphtory::vectors::{
-    cache::CachedEmbeddingModel, storage::LazyDiskVectorCache, template::DocumentTemplate,
-    vectorisable::Vectorisable, vectorised_graph::VectorisedGraph,
-};
 use raphtory::{
     db::{
         api::{
@@ -51,6 +44,15 @@ use std::{
 use tokio::sync::{OwnedRwLockReadGuard, OwnedRwLockWriteGuard, RwLock};
 use tracing::{error, warn};
 use walkdir::WalkDir;
+
+#[cfg(feature = "vectors")]
+use {
+    crate::model::graph::vectorised_graph::GqlVectorisedGraph,
+    raphtory::vectors::{
+        cache::CachedEmbeddingModel, storage::LazyDiskVectorCache, template::DocumentTemplate,
+        vectorisable::Vectorisable, vectorised_graph::VectorisedGraph,
+    },
+};
 
 #[derive(thiserror::Error, Debug)]
 pub enum ParquetPathError {
@@ -883,6 +885,7 @@ impl Data {
     /// Checks read permission then returns the vectorised graph, if any.
     /// Returns `None` for filtered-access users: embeddings are computed from the full graph
     /// and search results cannot be retroactively row-filtered.
+    #[cfg(feature = "vectors")]
     pub(crate) async fn get_vectors_with_read_permission(
         &self,
         ctx: &Context<'_>,
@@ -892,16 +895,8 @@ impl Data {
         if matches!(perm, GraphPermission::Read { filter: Some(_) }) {
             return Ok(None);
         }
-        #[cfg(feature = "vectors")]
-        {
-            let graph = self.get_graph(path).await?;
-            Ok(graph.vectors().cloned().map(|g| g.into()))
-        }
-        #[cfg(not(feature = "vectors"))]
-        {
-            let _ = path;
-            Err(async_graphql::Error::new("vectors feature not enabled"))
-        }
+        let graph = self.get_graph(path).await?;
+        Ok(graph.vectors().cloned().map(|g| g.into()))
     }
 }
